Add include file for depth_map_to_pointcloud
